Stuffs (Ai go to their seats)
AI go to their seat location and then plop their butt on the floor because I realized at 12:15 we didn't have an animation for sitting in a chair... but we DO have an animation for sitting on the ground.

Able to Spawn AI and then the go to a seat in the church.